testing

+++ example.c
@@ -17,7 +17,7 @@ int done;
 void handler1(int signum) {
     if (signum == SIGUSR1) {
         printf("\nGot SIGUSR1\n");
-        printf("this is handler99(): got a signal %d\n", signum);
+        printf("this is handler1(): got a signal %d\n", signum);
     }
     if (signum == SIGUSR2) {
         printf("\nGot SIGUSR2\n");
@@ -46,20 +46,20 @@ int main() {
     pid = fork();
     if (pid > 0) {
         printf("I am the parent, pid: %d\n", getpid());
-//        ptr = (char *) shmat(memid, 0, 0);
-//        if (ptr == NULL) {
-//            printf("shmat() failed\n");
-//            return(8);
-//        }
-//        printf("Parent is writing '%s' to the shared memory\n", buffer);
-//        strcpy(ptr, buffer);
-        //wait(NULL);
+        ptr = (char *) shmat(memid, 0, 0);
+        if (ptr == NULL) {
+            printf("shmat() failed\n");
+            return(8);
+        }
+        printf("Parent is writing '%s' to the shared memory\n", buffer);
+        strcpy(ptr, buffer);
+        wait(NULL);
         kill(getpid(), SIGUSR1);
     } else {
         printf("\nI am the child, pid: %d\n", getpid());
-//        ptr = (char *) shmat(memid, 0, 0);
-//        printf("I am the child, and I read this from the shared memory: '%s'\n", ptr);
-        //shmdt(ptr);
+        ptr = (char *) shmat(memid, 0, 0);
+        printf("I am the child, and I read this from the shared memory: '%s'\n", ptr);
+        shmdt(ptr);
         kill(getpid(), SIGUSR2);
     }
     while ( ! done );
